Fixed universal-hot-reload
